requirements-dev.txt
-r function/requirements.txt

black==22.1.0
flake8==4.0.1
pytest==7.0.1
aws-cdk-lib==2.15.0
boto3==1.21.13
moto[rekognition,s3]==4.0.1
